Move ServiceHelper to new header
